% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patest.R
\docType{data}
\name{keystone}
\alias{keystone}
\title{Keystone}
\format{
A data frame with 22,101 rows and 14 variables
\describe{
  \item{year}{School year (ending) the Keystone was administered}
  \item{county}{School district county}
  \item{district}{School district name}
  \item{school_number}{Identification for the school}
  \item{school_name}{Name of the school}
  \item{subject}{"Literature", "Algebra I", or "Biology"}
  \item{grade}{11 (only admistered in 11th grade)}
  \item{group}{Student group, "All Students" or "Historically Underperforming"}
  \item{students}{Number of students who took the assessment}
  \item{below_basic}{Proportion of students who scored below basic}
  \item{basic}{Proportion of students who scored basic}
  \item{proficient}{Proportion of students who scored proficient}
  \item{advanced}{Proportion of students who scored advanced}
}
}
\usage{
keystone
}
\description{
Data from Pennsyvlania's Keystone assessments
}
\keyword{datasets}
