% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patest.R
\docType{data}
\name{pa_act}
\alias{pa_act}
\title{PA ACT}
\format{
A data frame with 5,770 rows and 12 variables
\describe{
  \item{year}{School year (ending) the Keystone was administered}
  \item{district}{School district name}
  \item{school_number}{Identification for the school}
  \item{school_name}{Name of the school}
  \item{students}{Number of students who took the assessment}
  \item{english}{Average english score for school}
  \item{math}{Average math score for school}
  \item{reading}{Average reading score for school}
  \item{science}{Average science score for school}
  \item{writing}{Average writing score for school}
  \item{english_writing}{Average english and writing composite score for school}
  \item{composite}{Average composite score for school}
}
}
\usage{
pa_act
}
\description{
ACT results from PA public schools
}
\keyword{datasets}
