% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patest.R
\docType{data}
\name{pa_sat}
\alias{pa_sat}
\title{PA SAT}
\format{
A data frame with 5,502 rows and 11 variables
\describe{
  \item{year}{School year (ending) the Keystone was administered}
  \item{district}{School district name}
  \item{school_number}{Identification for the school}
  \item{school_name}{Name of the school}
  \item{students}{Number of students who took the assessment}
  \item{verbal}{Average verbal score for school}
  \item{math}{Average math score for school}
  \item{writing}{Average writing score for school}
  \item{reading}{Average reading score for school}
  \item{reading_writing}{Average reading and writing composite score for school}
  \item{composite}{Average composite score for school}
}
}
\usage{
pa_sat
}
\description{
SAT results from PA public schools from 2001 to 2019
}
\keyword{datasets}
