% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patest.R
\docType{package}
\name{patest}
\alias{patest}
\title{PA Test (Package)}
\description{
The Pennsylvania System of School Assessment, also known as PSSA, measures
  how well students have achieved in reading, mathematics, science and
  writing according to Pennsylvania's world-class academic standards. By
  using these standards, educators, parents and administrators can evaluate
  their students' strengths and weaknesses to increase students' achievement
  scores.
}
\details{
Pennsylvania school testing results
}
\references{
https://www.education.pa.gov/K-12/Assessment%20and%20Accountability/PSSA/Pages/default.aspx
}
